requirements.txt
Package             Version
------------------- --------------------
backcall            0.2.0
certifi             2022.12.7
charset-normalizer  3.1.0
colorama            0.4.6
contourpy           1.3.0
cycler              0.11.0
debugpy             1.6.7
decorator           5.1.1
easyocr             1.6.2
entrypoints         0.4
filelock            3.16.1
fonttools           4.38.0
fsspec              2024.12.0
gitdb               4.0.10
GitPython           3.1.31
idna                3.4
imageio             2.28.0
importlib_resources 6.4.5
imutils             0.5.4
ipykernel           6.16.2
ipython             7.34.0
jedi                0.18.2
Jinja2              3.1.5
jupyter_client      7.4.9
jupyter_core        4.12.0
kiwisolver          1.4.4
MarkupSafe          3.0.2
matplotlib          3.9.4
matplotlib-inline   0.1.6
mpmath              1.3.0
munkres             1.1.4
nest-asyncio        1.5.6
networkx            2.6.3
ninja               1.11.1
numpy               1.23.0
opencv-python       4.10.0.84
packaging           24.2
pandas              1.3.5
parso               0.8.3
pickleshare         0.7.5
Pillow              9.5.0
pip                 24.3.1
prompt-toolkit      3.0.38
psutil              5.9.5
pyclipper           1.3.0.post4
Pygments            2.15.1
pyparsing           3.0.9
PyQt5               5.15.9
PyQt5-Qt5           5.15.2
PyQt5-sip           12.12.1
PySide6             6.8.1
pytesseract         0.3.13
python-bidi         0.4.2
python-dateutil     2.8.2
pytz                2023.3
PyWavelets          1.3.0
pywin32             306
PyYAML              6.0
pyzmq               25.0.2
requests            2.29.0
scikit-image        0.19.3
scipy               1.11.2
seaborn             0.12.2
sentry-sdk          1.21.1
setuptools          75.6.0
shapely             2.0.1
shiboken6           6.8.1
six                 1.16.0
smmap               5.0.0
sympy               1.13.1
tesseract           0.1.3
thop                0.1.1.post2209072238
tifffile            2021.11.2
torch               2.5.1+cu118
torchaudio          2.5.1+cu118
torchvision         0.20.1+cu118
tornado             6.2
tqdm                4.65.0
traitlets           5.9.0
typing_extensions   4.9.0
tzdata              2024.2
ultralytics         8.0.89
unicodedata2        15.1.0
urllib3             1.26.15
wcwidth             0.2.6
wheel               0.45.1
zipp                3.21.0
